ETL_pipeline/pipeline/slack_bot/requirements.txt
pandas
pyjokes
sqlalchemy
psycopg2-binary
requests
python-dotenv